Restore search after TB hit
But this time with a depth depending on the distance
from root. If we are far from root, return quickly, if
we are near root do an half depth search.

Always return when in draw.

This should increase efficency in the middlegame and
avoid artifacts in the endgame.

Also introduce symmetry among winning and losing TB scores.